Move and break up RxLocationServices
