+++ gnd/src/main/java/com/google/android/gnd/system/LocationManager.java
@@ -26,9 +26,9 @@ import com.google.android.gms.location.LocationRequest;
 import com.google.android.gnd.inject.ActivityScoped;
 import com.google.android.gnd.model.feature.Point;
 import com.google.android.gnd.rx.BooleanOrError;
-import com.google.android.gnd.rx.RxLocationServices;
-import com.google.android.gnd.rx.RxLocationServices.RxFusedLocationProviderClient;
-import com.google.android.gnd.rx.RxLocationServices.RxFusedLocationProviderClient.RxLocationCallback;
+import com.google.android.gnd.system.rx.RxFusedLocationProviderClient;
+import com.google.android.gnd.system.rx.RxLocationCallback;
+import com.google.android.gnd.system.rx.RxLocationServices;
 import io.reactivex.BackpressureStrategy;
 import io.reactivex.Flowable;
 import io.reactivex.Maybe;
