+++ gnd/src/main/java/com/google/android/gnd/system/rx/RxLocationServices.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.gnd.system.rx;
+
+import android.content.Context;
+
+/** Thin wrapper around LocationServices exposing key features as reactive streams. */
+public class RxLocationServices {
+  public static RxSettingsClient getSettingsClient(Context context) {
+    return new RxSettingsClient(context);
+  }
+
+  public static RxFusedLocationProviderClient getFusedLocationProviderClient(Context context) {
+    return new RxFusedLocationProviderClient(context);
+  }
+}
